split pace.driver.run module into multiple files
